services/auth-middleware/requirements.txt
fastapi
PyJWT
uvicorn
python-dotenv
pydantic